+++ src/utilities.ts
@@ -1,2 +1,3 @@
 export { msToParts } from './msToParts.js';
 export { parseDuration } from './parseDuration.js';
+export { partsToMs } from './partsToMs.js';